Use yield instead of a list.
I don't want to rewrite the sync function to be one monolithic thing,
although it now would be more sensible than it was ever before. The
reason for this is that i want vdirsyncer to precompute all necessary
actions before it actually does anything, which probably helps in the
future with a dry-run mode.