Port CLI to Click.
After a while i finally gave in and ported the whole command line
interface of vdirsyncer to Click. While it might seem like a huge
dependency, i hope we can eliminate some helper functions in the
process, and maybe are more motivated to write a more beautiful and
intuitive interface due to all the niceties Click provides.