Add double quote in exemple config files (#732)
* nextcloud.rst : add double quote to not forget them

Add double quote to not forget them, and avoid the message :
warning: Soon, all strings have to be in double quotes. Please replace UserName with "UserName"

* fastmail.rst : add double quote to not forget them

Add double quote to not forget them, and avoid the message :
warning: Soon, all strings have to be in double quotes. Please replace UserName with "UserName"

* icloud.rst : add double quote to not forget them

Add double quote to not forget them, and avoid the message :
warning: Soon, all strings have to be in double quotes. Please replace UserName with "UserName"

* todoman.rst : add double quote to not forget them

Add double quote to not forget them, and avoid the message :
warning: Soon, all strings have to be in double quotes. Please replace UserName with "UserName"

* xandikos.rst : add double quote to not forget them

Add double quote to not forget them, and avoid the message :
warning: Soon, all strings have to be in double quotes. Please replace UserName with "UserName"

* davmail.rst : add double quote to not forget them

Add double quote to not forget them, and avoid the message :
warning: Soon, all strings have to be in double quotes. Please replace UserName with "UserName"

* partial-sync.rst : add double quote to not forget them

Add double quote to not forget them, and avoid the message :
warning: Soon, all strings have to be in double quotes. Please replace UserName with "UserName"