storage/http: expose `ignore_uids` config parameter
This allows users to choose whether or not to ignore the UIDs coming
from WebCAL providers and instead use the item's content hash. This
remains the default behaviour.

In my case, my provider was generating stable UIDs but was patching the
export date in other fields, resulting in the hash of all events
changing at each synchronisation, having the opposite effect.